Heart Disease Prediction System
 the project objective is to detect whether patients have heart disease or not be given a number of features from patients. The motivation of our project is to save human resources in medical centers and improve the accuracy of diagnosis. In our project, we used different methods to visualize data and trained a machine learning model to detect heart disease by using the Logistic Regression model.
Limitations 
Medical diagnosis is considered a significant yet intricate task that needs to be carried out
precisely and efficiently. The automation of the same would be highly beneficial. Clinical
decisions are often made based on the doctor’s intuition and experience rather than on the
knowledge-rich data hidden in the database. This practice leads to unwanted biases, errors, and
excessive medical costs which affects the quality of service provided to patients

CONCLUSION / FUTURE WORK
We use some libraries provided by Python to implement this project. After creating the model, the algorithm of Logistic Regression gives us the test accuracy, which is 81.96%. Though we get a good result of 81.96% accuracy, that is not enough because it cannot guarantee that no wrong diagnosis happens. To improve accuracy, we hope to require more datasets. There are many possible improvements that could be explored to improve the scalability and accuracy of this prediction system. In the future, to predict disease we want to try different diseases such as lung cancer and want to apply more models to this project to gain more accuracy.